smart_notebook_be/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.1
openai==1.10.0
pydantic==2.6.0
